Add files via upload
The Caesar Cipher is one of the oldest and most famous encryption techniques. It works by shifting each letter in the message by a fixed number of positions in the alphabet, creating a simple yet effective way to conceal the original text. For example with a shift of 1, A would be replaced by B, B would become C, and so on.  This code is written in java .